requirements.txt
Adafruit-Blinka==8.69.0
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-ht16k33==4.6.15
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.87.0
Adafruit-PureIO==1.1.11
Adafruit_GPIO==1.0.3
Adafruit_LED_Backpack==1.8.1
binho-host-adapter==0.1.6
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
pillow==12.1.0
pyftdi==0.57.1
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.3.1
requests==2.32.5
RPi.GPIO==0.7.1
six==1.17.0
spidev==3.8
sysv_ipc==1.2.0
toml==0.10.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
